+++ lib/rubocop/cop/lint/percent_string_array.rb
@@ -40,7 +40,6 @@ module RuboCop
           add_offense(node)
         end
 
-        # rubocop:disable Performance/HashEachMethods
         def autocorrect(node)
           lambda do |corrector|
             node.values.each do |value|
@@ -55,7 +54,6 @@ module RuboCop
             end
           end
         end
-        # rubocop:enable Performance/HashEachMethods
 
         private